// src/app/ws/soccer/model/AccountStatisticDTO.ts
/**
 * Soccer API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * Contact: voetbalsvk@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import * as models from "./models";

export interface AccountStatisticDTO {
    account?: models.AccountDTO;

    assists?: number;

    goals?: number;

    played?: number;

}